Improved perf tests and geo exceptions

+++ tools/test_utils/include/TestUtils/MakeMeshes.hpp
@@ -57,8 +57,8 @@ std::shared_ptr<meshkernel::Mesh2D> MakeRectangularMeshForTesting(
     meshkernel::Point const& origin = {0.0, 0.0});
 
 std::shared_ptr<meshkernel::Mesh2D> MakeRectangularMeshForTesting(
-    int n,
-    int m,
+    size_t n,
+    size_t m,
     double delta,
     meshkernel::Projection projection,
     meshkernel::Point const& origin = {0.0, 0.0});
